change code to meet bioconductor's requirements

+++ R/align_free_quan.R
@@ -6,28 +6,29 @@ convert_data <- function()
     list.files(pattern = "quant.sf",
                recursive = TRUE,
                full.names = TRUE)
-
-  others <-
-    data.frame(
-      transcript_id = character(),
-      length = numeric(),
-      TPM = numeric(),
-      count = numeric(),
-      sample = character()
-    )
-  for (f in files)
-  {
-    other <- utils::read.table(f, header = TRUE, sep = "\t")
-    other$sample <- gsub("^\\./", "", gsub("_trans.*", "", f))
-    others <- rbind(others, other[, -2])
+  if(length(files)){
+    others <-
+      data.frame(
+        transcript_id = character(),
+        length = numeric(),
+        TPM = numeric(),
+        count = numeric(),
+        sample = character()
+      )
+    for (f in files)
+    {
+      other <- utils::read.table(f, header = TRUE, sep = "\t")
+      other$sample <- gsub("^\\./", "", gsub("_trans.*", "", f))
+      others <- rbind(others, other[, -2])
+    }
+    names(others) <-
+      c("transcript_id", "length", "TPM", "count", "sample")
+    others <-
+      others[, c("sample", "transcript_id", "count", "TPM", "length")]
+    utils::write.csv(others,
+                    "transcript_alignment_free_quantification.csv",
+                    row.names = FALSE)
   }
-  names(others) <-
-    c("transcript_id", "length", "TPM", "count", "sample")
-  others <-
-    others[, c("sample", "transcript_id", "count", "TPM", "length")]
-  utils::write.csv(others,
-                   "salmon_transcript_quantifications.csv",
-                   row.names = FALSE)
 }
 
 tx2gene <- function()
@@ -37,49 +38,53 @@ tx2gene <- function()
                recursive = TRUE,
                full.names = TRUE)
   #cmd1 <- paste("egrep -v '^#|^$'", annotation, "| awk -F '\t' '$3 ~ /RNA/ {print $9}' | awk -F ';' 'BEGIN{OFS = \"=\";} {print $1, $2;}' | awk -F '=' 'BEGIN{OFS = \"-\";}{print $NF, $2;}'| grep 'gene' | awk -F '-' 'BEGIN{OFS = \",\";print \"gene_id\", \"transcript_id\"}{print $2, $4}' > tx2gene.csv")
-  cmd1 <-
-    paste(
-      "egrep -v '^#|^$'",
-      annotation,
-      "| cut -f 9 | grep ID=rna | awk -F ';' 'BEGIN{OFS = \"=\";} {print $1, $2;}' | awk -F '=' 'BEGIN{OFS = \",\"} {print $NF, $2}' > raw_tx2gene.csv"
-    )
-  #cat(cmd1)
-  system(cmd1)
-  tx2gene <- utils::read.csv("raw_tx2gene.csv", header = FALSE)
-  tx2gene[] <- lapply(tx2gene, as.character)
-  index_to_be_changed <- which(tx2gene[, 1] %in% tx2gene[, 2])
-  b <- length(index_to_be_changed)
-  if (b > 0) {
-    for (i in seq_len(b))
-    {
-      tx2gene[index_to_be_changed[i], 1] <-
-        tx2gene[tx2gene[, 2] == tx2gene[index_to_be_changed[i], 1] , 1]
+  if(length(annotation)){
+    cmd1 <-
+      paste(
+        "egrep -v '^#|^$'",
+        annotation,
+        "| cut -f 9 | grep ID=rna | awk -F ';' 'BEGIN{OFS = \"=\";} {print $1, $2;}' | awk -F '=' 'BEGIN{OFS = \",\"} {print $NF, $2}' > raw_tx2gene.csv"
+      )
+    #cat(cmd1)
+    system(cmd1)
+    tx2gene <- utils::read.csv("raw_tx2gene.csv", header = FALSE)
+    tx2gene[] <- lapply(tx2gene, as.character)
+    index_to_be_changed <- which(tx2gene[, 1] %in% tx2gene[, 2])
+    b <- length(index_to_be_changed)
+    if (b > 0) {
+      for (i in seq_len(b))
+      {
+        tx2gene[index_to_be_changed[i], 1] <-
+          tx2gene[tx2gene[, 2] == tx2gene[index_to_be_changed[i], 1] , 1]
+      }
     }
+    tx2gene[, 1] <- gsub("gene-", "", tx2gene[, 1])
+    tx2gene[, 2] <- gsub("rna-", "", tx2gene[, 2])
+    tx2gene <- tx2gene[, c(2, 1)]
+    colnames(tx2gene) <- c("transcript_id", "gene_id")
+    utils::write.csv(tx2gene, row.names = FALSE, "tx2gene.csv")
   }
-  tx2gene[, 1] <- gsub("gene-", "", tx2gene[, 1])
-  tx2gene[, 2] <- gsub("rna-", "", tx2gene[, 2])
-  tx2gene <- tx2gene[, c(2, 1)]
-  colnames(tx2gene) <- c("transcript_id", "gene_id")
-  utils::write.csv(tx2gene, row.names = FALSE, "tx2gene.csv")
 }
 
 utils::globalVariables("TPM")
 
 gene_quan <- function()
 {
-  tx2gene_file <- utils::read.csv("tx2gene.csv")
-  transcript <-
-    utils::read.csv("salmon_transcript_quantifications.csv")
-  #transcript <- transcript[, -1]
-  all <-
-    merge(tx2gene_file, transcript, by = "transcript_id", all = TRUE)
-  all <- stats::na.omit(all)
-  tmp1 <-
-    all %>% dplyr::group_by(sample, gene_id) %>% dplyr::summarise(count = sum(count))
-  gene_quantification <- tmp1[, c("sample", "gene_id", "count")]
-  utils::write.csv(gene_quantification,
-                   "salmon_gene_quantification.csv",
-                   row.names = FALSE)
+  if(file.exists ("tx2gene.csv")&&file.exists("transcript_alignment_free_quantification.csv")){
+    tx2gene_file <- utils::read.csv("tx2gene.csv")
+    transcript <-
+      utils::read.csv("transcript_alignment_free_quantification.csv")
+    #transcript <- transcript[, -1]
+    all <-
+      merge(tx2gene_file, transcript, by = "transcript_id", all = TRUE)
+    all <- stats::na.omit(all)
+    tmp1 <-
+      all %>% dplyr::group_by(sample, gene_id) %>% dplyr::summarise(count = sum(count))
+    gene_quantification <- tmp1[, c("sample", "gene_id", "count")]
+    utils::write.csv(gene_quantification,
+                    "gene_alignment_free_quantification.csv",
+                    row.names = FALSE)
+  } 
 }
 
 #' Alignment-free expression quantification with Salmon at gene and transcript levels.
@@ -92,75 +97,80 @@ gene_quan <- function()
 #' @export align_free_quan
 #' @return None
 #' @examples
-#' \dontrun{
 #'
-#' align_free_quan("paired", "sesame.fna", "transcript_sesame.fna","sesame.gff")
-#'}
+#' align_free_quan(pair = "paired", "test_Homo_sapiens.fna", "test_transcript_Homo_sapiens.fna","test_Homo_sapiens.gff")
+#'
+#'
 #'
 align_free_quan <- function(pair, genome, transcript, annotation)
 {
-  cmd3 <- paste("salmon index -t", transcript, "-i salmon_index")
-  # cat(cmd3, "\n")
-  system(cmd3)
-  if (pair == "paired")
-  {
-    read <-
-      list.files(pattern = "^Trimmed.*1\\.fastq$", full.names = FALSE)
-    if (length(read) == 0) {
-      read <- list.files(pattern = ".*1\\.fastq$", full.names = FALSE)
-    }
-    for (f in read)
+  if(file.exists(genome)&&file.exists(transcript)&&file.exists(annotation)){
+    cmd3 <- paste("salmon index -t", transcript, "-i salmon_index")
+    # cat(cmd3, "\n")
+    system(cmd3)
+    if (pair == "paired")
     {
-      # read1 <- paste(read1, sep = ",", collapse = ',')
-      # read2 <- paste(read2, sep = ",", collapse = ',')
-      name <- gsub("_1.fastq", "", f)
-      out <- paste0(name, "_transcripts_quant")
-      read1 <- paste0(name, "_1.fastq")
-      read2 <- paste0(name, "_2.fastq")
-      # cmd4 <- paste("salmon quant -i salmon_index -l A", gentrome.fna, "-1", read1, "-2", read2, "--validateMappings -o", out)
-      cmd4 <-
-        paste(
-          "salmon quant -i salmon_index -l A",
-          "-1",
-          read1,
-          "-2",
-          read2,
-          "--validateMappings -o",
-          out
-        )
-      # cat(cmd4, "\n")
-      system(cmd4, intern = TRUE)
-    }
-  } else if (pair == "single") {
-    read <-
-      list.files(pattern = "^Trimmed.*\\.fastq$", full.names = FALSE)
-    if (length(read) == 0) {
-      read <- list.files(pattern = ".*\\.fastq$", full.names = FALSE)
-    }
-    # read <- paste(read, sep = ",", collapse = ',')
-    for (f in read)
-    {
-      name <- gsub(".fastq", "", f)
-      out <- paste0(name, "_transcripts_quant")
-      # cmd4 <- paste("salmon quant -i salmon_index -l A -1", gentrome.fna, "-r", f, "--validateMappings -o", out)
-      cmd4 <-
-        paste("salmon quant -i salmon_index -l A -r",
-              f,
+      read <-
+        list.files(pattern = "^Trimmed.*1\\.fastq$", full.names = FALSE)
+      if (length(read) == 0) {
+        read <- list.files(pattern = ".*1\\.fastq$", full.names = FALSE)
+      }
+      if(length(read)){
+        for (f in read)
+        {
+          # read1 <- paste(read1, sep = ",", collapse = ',')
+          # read2 <- paste(read2, sep = ",", collapse = ',')
+          name <- gsub("_1.fastq", "", f)
+          out <- paste0(name, "_transcripts_quant")
+          read1 <- paste0(name, "_1.fastq")
+          read2 <- paste0(name, "_2.fastq")
+          # cmd4 <- paste("salmon quant -i salmon_index -l A", gentrome.fna, "-1", read1, "-2", read2, "--validateMappings -o", out)
+          cmd4 <-
+            paste(
+              "salmon quant -i salmon_index -l A",
+              "-1",
+              read1,
+              "-2",
+              read2,
               "--validateMappings -o",
-              out)
+              out
+            )
+          # cat(cmd4, "\n")
+          system(cmd4, intern = TRUE)
+        }
+      } else print("No fastq files are found in the work directory.")
+    } else if (pair == "single") {
+      read <-
+        list.files(pattern = "^Trimmed.*\\.fastq$", full.names = FALSE)
+      if (length(read) == 0) {
+        read <- list.files(pattern = ".*\\.fastq$", full.names = FALSE)
+      }
+      if(length(read)){
+      # read <- paste(read, sep = ",", collapse = ',')
+        for (f in read)
+        {
+          name <- gsub(".fastq", "", f)
+          out <- paste0(name, "_transcripts_quant")
+          # cmd4 <- paste("salmon quant -i salmon_index -l A -1", gentrome.fna, "-r", f, "--validateMappings -o", out)
+          cmd4 <-
+            paste("salmon quant -i salmon_index -l A -r",
+                  f,
+                  "--validateMappings -o",
+                  out)
 
-      # cat(cmd4, "\n")
-      system(cmd4, intern = TRUE)
-    }
-
-  } else
-    stop("Paired-end and single-end mix. Please check the data source!")
-  convert_data() #### rna quantification to use for quantifying genes.
-  tx2gene()
-  gene_quan()
-  unlink("raw_tx2gene.csv")
-  unlink("tx2gene.csv")
-  folders <- dir(pattern = "transcripts_quant$")
-  unlink(folders, recursive = TRUE)
-  unlink("salmon_index", recursive = TRUE)
+          # cat(cmd4, "\n")
+          system(cmd4, intern = TRUE)
+        }        
+      }
+    } else
+      stop("Paired-end and single-end mix. Please check the data source!")
+    convert_data() #### rna quantification to use for quantifying genes.
+    tx2gene()
+    gene_quan()
+    unlink("raw_tx2gene.csv")
+    unlink("tx2gene.csv")
+    folders <- dir(pattern = "transcripts_quant$")
+    unlink(folders, recursive = TRUE)
+    unlink("salmon_index", recursive = TRUE)
+  } else print("No reference genome and annotation files are found. Please download them first")
 }
